To use the ESP8266 as a Wi-Fi station and access point simultaneously, you can use the Arduino IDE and the ESP8266WiFi library. Below is an example of how you can set up the ESP8266 to act as both a Wi-Fi station (to connect to an existing Wi-Fi network) and an access point (to create its own Wi-Fi network)
This code will first connect the ESP8266 to the specified Wi-Fi network as a station using the provided SSID and password. Then, it will create an access point with the given SSID and password. The ESP8266 will act as a gateway between devices connected to its access point and the internet through the station connection.

Please make sure to replace "YourWiFiSSID" and "YourWiFiPassword" with your actual Wi-Fi network's SSID and password. Also, remember that when the ESP8266 is acting as an access point, it won't have access to the internet directly. If you want it to have internet access even in access point mode, you can set up additional configurations to allow the ESP8266 to forward traffic between its station and access point interfaces.